+++ backend/src/main/java/feedzupzup/backend/feedback/controller/AdminFeedbackController.java
@@ -3,7 +3,7 @@ package feedzupzup.backend.feedback.controller;
 import feedzupzup.backend.admin.dto.AdminSession;
 import feedzupzup.backend.feedback.api.AdminFeedbackApi;
 import feedzupzup.backend.feedback.application.AdminFeedbackService;
-import feedzupzup.backend.feedback.domain.vo.FeedbackSortBy;
+import feedzupzup.backend.feedback.domain.vo.FeedbackSortType;
 import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
 import feedzupzup.backend.feedback.dto.request.UpdateFeedbackCommentRequest;
 import feedzupzup.backend.feedback.dto.response.AdminFeedbackListResponse;
@@ -29,7 +29,7 @@ public class AdminFeedbackController implements AdminFeedbackApi {
             final int size,
             final Long cursorId,
             final ProcessStatus status,
-            final FeedbackSortBy sortBy
+            final FeedbackSortType sortBy
     ) {
         return SuccessResponse.success(HttpStatus.OK, adminFeedbackService.getFeedbackPage(
                 loginOrganizerInfo.organizationUuid(),
