+++ backend/src/main/java/feedzupzup/backend/feedback/domain/service/sort/FeedbackSortStrategy.java
@@ -1,7 +1,7 @@
 package feedzupzup.backend.feedback.domain.service.sort;
 
 import feedzupzup.backend.feedback.domain.Feedback;
-import feedzupzup.backend.feedback.domain.vo.FeedbackSortBy;
+import feedzupzup.backend.feedback.domain.vo.FeedbackSortType;
 import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
 import feedzupzup.backend.feedback.dto.response.FeedbackItem;
 import java.util.List;
@@ -13,7 +13,7 @@ public interface FeedbackSortStrategy {
     List<FeedbackItem> getSortedFeedbacks(final UUID organizationUuId, final ProcessStatus status, final Long cursorId,
             final Pageable pageable);
 
-    FeedbackSortBy getType();
+    FeedbackSortType getType();
 
     List<FeedbackItem> sort(List<Feedback> feedbacks);
