+++ backend/src/main/java/feedzupzup/backend/feedback/domain/service/sort/FeedbackSortStrategyFactory.java
@@ -2,7 +2,7 @@ package feedzupzup.backend.feedback.domain.service.sort;
 
 import static java.util.stream.Collectors.toMap;
 
-import feedzupzup.backend.feedback.domain.vo.FeedbackSortBy;
+import feedzupzup.backend.feedback.domain.vo.FeedbackSortType;
 import java.util.List;
 import java.util.Map;
 import org.springframework.stereotype.Component;
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class FeedbackSortStrategyFactory {
 
-    private final Map<FeedbackSortBy, FeedbackSortStrategy> strategies;
+    private final Map<FeedbackSortType, FeedbackSortStrategy> strategies;
 
     public FeedbackSortStrategyFactory(final List<FeedbackSortStrategy> sortStrategies) {
         this.strategies = sortStrategies.stream()
@@ -20,7 +20,7 @@ public class FeedbackSortStrategyFactory {
                 ));
     }
 
-    public FeedbackSortStrategy find(final FeedbackSortBy sortBy) {
+    public FeedbackSortStrategy find(final FeedbackSortType sortBy) {
         if (!strategies.containsKey(sortBy)) {
             throw new IllegalArgumentException("해당 정렬(sortBy=" + sortBy + ")은 지원하지 않고 있습니다.");
         }
